Create authentication system and basic file structure
This commit includes the creation of fundamental files and directories as well as the implementation of an authentication system. It adds user signup, login, and token refresh functionalities using JWT (JSON Web Tokens) and golang fiber for routing. The database storage uses gorm with MySQL, which is configured in the .env file. The user and JWT interfacing is defined in the domain repositories and external use cases, respectively. Added .air.toml file for the air live reloading tool configuration. Also, the essential dependency packages are included in go.mod file.